+++ programs/coreData.py
@@ -1,6 +1,12 @@
 #!/usr/bin/env python
 # coding: utf-8
 
+# In[ ]:
+
+
+
+
+
 # <img align="right" src="images/dans-small.png"/>
 # <img align="right" src="images/tf-small.png"/>
 # <img align="right" src="images/etcbc.png"/>
@@ -38,7 +44,7 @@
 # ```sudo -H pip3 install --upgrade text-fabric
 # ```
 
-# In[1]:
+# In[19]:
 
 
 import os
@@ -66,6 +72,9 @@ if "SCRIPT" not in locals():
     )
 
 
+# In[3]:
+
+
 def stop(good=False):
     if SCRIPT:
         sys.exit(0 if good else 1)
@@ -76,19 +85,31 @@ def stop(good=False):
 # The conversion is executed in an environment of directories, so that sources, temp files and
 # results are in convenient places and do not have to be shifted around.
 
-# In[3]:
+# In[4]:
 
 
 repoBase = os.path.expanduser("~/github/etcbc")
 thisRepo = "{}/{}".format(repoBase, CORE_NAME)
 
+
+# In[5]:
+
+
 thisSource = "{}/source/{}".format(thisRepo, VERSION)
 mqlzFile = "{}/{}.mql.bz2".format(thisSource, CORE_NAME)
+
+
+# In[6]:
+
 
 thisTemp = "{}/_temp/{}".format(thisRepo, VERSION)
 thisTempSource = "{}/source".format(thisTemp)
 mqlFile = "{}/{}.mql".format(thisTempSource, CORE_NAME)
 thisTempTf = "{}/tf".format(thisTemp)
+
+
+# In[7]:
+
 
 thisTf = "{}/tf/{}".format(thisRepo, VERSION)
 
@@ -98,7 +119,7 @@ thisTf = "{}/tf/{}".format(thisRepo, VERSION)
 # Check whether this conversion is needed in the first place.
 # Only when run as a script.
 
-# In[4]:
+# In[8]:
 
 
 if SCRIPT:
@@ -125,10 +146,14 @@ if SCRIPT:
 # We save the configs we need per source and version.
 # And we define a stripped down default version to start with.
 
-# In[5]:
+# In[9]:
 
 
 slotType = "word"
+
+
+# In[10]:
+
 
 featureMetaData = dict(
     dataset="BHSA",
@@ -139,6 +164,10 @@ featureMetaData = dict(
     website="https://shebanq.ancient-data.org",
     email="shebanq@ancient-data.org",
 )
+
+
+# In[11]:
+
 
 oText = {
     "": {
@@ -254,10 +283,14 @@ oText = {
 # The next function selects the proper otext material, falling back on a default if nothing
 # appropriate has been specified in `oText`.
 
-# In[6]:
+# In[12]:
 
 
 thisOtext = oText.get(VERSION, oText[""])
+
+
+# In[13]:
+
 
 if thisOtext is oText[""]:
     utils.caption(
@@ -287,15 +320,23 @@ else:
 # 
 # Check the source, utils.bunzip it if needed, empty the result directory.
 
-# In[7]:
+# In[14]:
 
 
 if not os.path.exists(thisTempSource):
     os.makedirs(thisTempSource)
 
+
+# In[15]:
+
+
 utils.caption(0, "bunzipping {} ...".format(mqlzFile))
 utils.bunzip(mqlzFile, mqlFile)
 utils.caption(0, "Done")
+
+
+# In[16]:
+
 
 if os.path.exists(thisTempTf):
     rmtree(thisTempTf)
@@ -306,7 +347,7 @@ os.makedirs(thisTempTf)
 # Transform the collected information in feature-like data-structures, and write it all
 # out to `.tf` files.
 
-# In[8]:
+# In[17]:
 
 
 TF = Fabric(locations=thisTempTf, silent=True)
@@ -316,7 +357,7 @@ TF.importMQL(mqlFile, slotType=slotType, otext=otextInfo, meta=featureMetaData)
 # # Rename features
 # We rename the features mentioned in the RENAME dictionary.
 
-# In[9]:
+# In[18]:
 
 
 if RENAME is None:
@@ -351,7 +392,7 @@ else:
 # For each changed feature we show the first line where the new feature differs from the old one.
 # We ignore changes in the metadata, because the timestamp in the metadata will always change.
 
-# In[10]:
+# In[19]:
 
 
 utils.checkDiffs(thisTempTf, thisTf)
@@ -361,7 +402,7 @@ utils.checkDiffs(thisTempTf, thisTf)
 # 
 # Copy the new TF dataset from the temporary location where it has been created to its final destination.
 
-# In[11]:
+# In[20]:
 
 
 utils.deliverDataset(thisTempTf, thisTf)
@@ -379,12 +420,16 @@ utils.deliverDataset(thisTempTf, thisTf)
 # At that point we have access to the full list of features.
 # We grab them and are going to load them all!
 
-# In[12]:
+# In[14]:
 
 
 utils.caption(4, "Load and compile standard TF features")
 TF = Fabric(locations=thisTf, modules=[""])
 api = TF.load("")
+
+
+# In[15]:
+
 
 utils.caption(4, "Load and compile all other TF features")
 allFeatures = TF.explore(silent=False, show=True)
@@ -395,7 +440,7 @@ api.makeAvailableIn(globals())
 
 # # Examples
 
-# In[13]:
+# In[17]:
 
 
 utils.caption(4, "Basic test")
@@ -405,7 +450,7 @@ for fmt in T.formats:
     utils.caption(0, "\t{}".format(T.text(range(1, 12), fmt=fmt)), continuation=True)
 
 
-# In[14]:
+# In[18]:
 
 
 if SCRIPT: