+++ tf/2021/freq_lex.tf
@@ -9,7 +9,7 @@
 @version=2021
 @website=https://shebanq.ancient-data.org
 @writtenBy=Text-Fabric
-@dateWritten=2021-08-24T11:15:19Z
+@dateWritten=2021-11-30T15:32:18Z
 
 15542
 51