+++ tf/2021/lex.tf
@@ -8,7 +8,7 @@
 @version=2021
 @website=https://shebanq.ancient-data.org
 @writtenBy=Text-Fabric
-@dateWritten=2021-08-24T11:13:11Z
+@dateWritten=2021-11-30T15:28:50Z
 
 B
 R>CJT/