+++ tf/2021/otext.tf
@@ -2,7 +2,7 @@
 @author=Eep Talstra Centre for Bible and Computer
 @dataset=BHSA
 @datasetName=Biblia Hebraica Stuttgartensia Amstelodamensis
-@dateWritten=2021-08-24T11:13:18Z
+@dateWritten=2021-11-30T15:28:55Z
 @email=shebanq@ancient-data.org
 @encoders=Constantijn Sikkel (QDF), Ulrik Petersen (MQL) and Dirk Roorda (TF)
 @fmt:lex-default={voc_lex_utf8} 
@@ -22,5 +22,5 @@
 @website=https://shebanq.ancient-data.org
 @writtenBy=Text-Fabric
 @writtenBy=Text-Fabric
-@dateWritten=2021-08-24T11:14:30Z
+@dateWritten=2021-11-30T15:31:02Z