+++ tf/2021/pargr.tf
@@ -8,7 +8,7 @@
 @version=2021
 @website=https://shebanq.ancient-data.org
 @writtenBy=Text-Fabric
-@dateWritten=2021-08-24T11:14:22Z
+@dateWritten=2021-11-30T15:30:23Z
 
 515690	1
 1