+++ tf/2021saved/book@el.tf
@@ -1,0 +1,55 @@
+@node
+@author=Eep Talstra Centre for Bible and Computer
+@dataset=BHSA
+@datasetName=Biblia Hebraica Stuttgartensia Amstelodamensis
+@email=shebanq@ancient-data.org
+@encoders=Dirk Roorda (TF)
+@language=Ελληνικά
+@languageCode=el
+@languageEnglish=greek
+@provenance=book names from wikipedia and other sources
+@valueType=str
+@version=2021
+@website=https://shebanq.ancient-data.org
+@writtenBy=Text-Fabric
+@dateWritten=2021-08-24T11:12:50Z
+
+426591	Γένεση
+Έξοδος
+Λευιτικό
+Αριθμοί
+Δευτερονόμιο
+Ιησούς
+Κριταί
+Σαμουήλ_A'
+Σαμουήλ_Β'
+Βασιλείς_A'
+Βασιλείς_Β'
+Ησαΐας
+Ιερεμίας
+Ιεζεκιήλ
+Ωσηέ
+Ιωήλ
+Αμώς
+Οβδιού
+Ιωνάς
+Μιχαίας
+Ναούμ
+Αβακκούμ
+Σοφονίας
+Αγγαίος
+Ζαχαρίας
+Μαλαχίας
+Ψαλμοί
+Ιώβ
+Παροιμίαι
+Ρουθ
+Άσμα_Ασμάτων
+Εκκλησιαστής
+Θρήνοι
+Εσθήρ
+Δανιήλ
+Έσδρας
+Νεεμίας
+Χρονικά_Α'
+Χρονικά_Β'