+++ tf/2021saved/book@he.tf
@@ -1,0 +1,55 @@
+@node
+@author=Eep Talstra Centre for Bible and Computer
+@dataset=BHSA
+@datasetName=Biblia Hebraica Stuttgartensia Amstelodamensis
+@email=shebanq@ancient-data.org
+@encoders=Dirk Roorda (TF)
+@language=עברית
+@languageCode=he
+@languageEnglish=hebrew
+@provenance=book names from wikipedia and other sources
+@valueType=str
+@version=2021
+@website=https://shebanq.ancient-data.org
+@writtenBy=Text-Fabric
+@dateWritten=2021-08-24T11:12:50Z
+
+426591	בראשית
+שמות
+ויקרא
+במדבר
+דברים
+יהושע
+שופטים
+שמואל_א
+שמואל_ב
+מלכים_א
+מלכים_ב
+ישעיהו
+ירמיהו
+יחזקאל
+הושע
+יואל
+עמוס
+עובדיה
+יונה
+מיכה
+נחום
+חבקוק
+צפניה
+חגי
+זכריה
+מלאכי
+תהילים
+איוב
+משלי
+רות
+שיר_השירים
+קהלת
+איכה
+אסתר
+דניאל
+עזרא
+נחמיה
+דברי_הימים_א
+דברי_הימים_ב