+++ tf/2021saved/book@hi.tf
@@ -1,0 +1,55 @@
+@node
+@author=Eep Talstra Centre for Bible and Computer
+@dataset=BHSA
+@datasetName=Biblia Hebraica Stuttgartensia Amstelodamensis
+@email=shebanq@ancient-data.org
+@encoders=Dirk Roorda (TF)
+@language=हिन्दी
+@languageCode=hi
+@languageEnglish=hindi
+@provenance=book names from wikipedia and other sources
+@valueType=str
+@version=2021
+@website=https://shebanq.ancient-data.org
+@writtenBy=Text-Fabric
+@dateWritten=2021-08-24T11:12:50Z
+
+426591	उत्पाति
+निर्गमन
+लैव्यव्यवस्थ
+गिनती
+व्यवस्थाविवरण
+यहोशू
+न्यायियों
+1_शमूएल
+2_शमूएल
+1_राजाओं
+2_राजाओं
+यशायाह
+यिर्मयाह
+यहेजकेल
+होशे
+योएल
+आमोस
+ओबधाह
+योना
+मीका
+नहूम
+हबक्कूक
+सपन्याह
+हाग्गै
+जकर्याह
+मलाकी
+भजन_संहिता
+अय्यूब
+नीतिबचन
+रुत
+श्रेष्ठगीत
+सभोपदेशक
+विलापगेत
+एस्तेर
+दानिय्यल
+एज्रा
+नहेम्याह
+1_इतिहास
+2_इतिहास