+++ tf/2021saved/book@ko.tf
@@ -1,0 +1,55 @@
+@node
+@author=Eep Talstra Centre for Bible and Computer
+@dataset=BHSA
+@datasetName=Biblia Hebraica Stuttgartensia Amstelodamensis
+@email=shebanq@ancient-data.org
+@encoders=Dirk Roorda (TF)
+@language=한국어
+@languageCode=ko
+@languageEnglish=korean
+@provenance=book names from wikipedia and other sources
+@valueType=str
+@version=2021
+@website=https://shebanq.ancient-data.org
+@writtenBy=Text-Fabric
+@dateWritten=2021-08-24T11:12:50Z
+
+426591	창세기
+출애굽기
+레위기
+민수기
+신명기
+여호수아
+사사기
+사무엘상
+사무엘하
+열왕기상
+열왕기하
+이사야
+예레미야
+에스겔
+호세아
+요엘
+아모스
+오바댜
+요나
+미가
+나훔
+하박국
+스바냐
+학개
+스가랴
+말라기
+시편
+욥기
+잠언
+룻기
+아가
+전도서
+예레미야애가
+에스더
+다니엘
+에스라
+느헤미야
+역대상
+역대하