+++ tf/2021saved/book@pa.tf
@@ -1,0 +1,55 @@
+@node
+@author=Eep Talstra Centre for Bible and Computer
+@dataset=BHSA
+@datasetName=Biblia Hebraica Stuttgartensia Amstelodamensis
+@email=shebanq@ancient-data.org
+@encoders=Dirk Roorda (TF)
+@language=ਪੰਜਾਬੀ
+@languageCode=pa
+@languageEnglish=punjabi
+@provenance=book names from wikipedia and other sources
+@valueType=str
+@version=2021
+@website=https://shebanq.ancient-data.org
+@writtenBy=Text-Fabric
+@dateWritten=2021-08-24T11:12:50Z
+
+426591	ਉਤਪਤ
+ਕੂਚ
+ਲੇਵੀਆਂ_ਦੀ_ਪੋਥੀ
+ਗਿਣਤੀ
+ਬਿਵਸਥਾ_ਸਾਰ
+ਯਹੋਸ਼ੁਆ
+ਨਿਆਂਈਆਂ_ਦੀ_ਪੋਥੀ
+1_ਸਮੂਏਲ
+2_ਸਮੂਏਲ
+1_ਰਾਜਿਆਂ
+2_ਰਾਜਿਆਂ
+ਯਸਾਯਾਹ
+ਯਿਰਮਿਯਾਹ
+ਹਿਜ਼ਕੀਏਲ
+ਹੋਸ਼ੇਆ
+ਯੋਏਲ
+ਆਮੋਸ
+ਓਬਦਯਾਹ
+ਯੂਨਾਹ
+ਮੀਕਾਹ
+ਨਹੂਮ
+ਹਬਕੱੂਕ
+ਸਫ਼ਨਯਾਹ
+ਹੱਜਈ
+ਜ਼ਕਰਯਾਹ
+ਮਲਾਕੀ
+ਜ਼ਬੂਰ
+ਅੱਯੂਬ
+ਕਹਾਉਤਾਂ
+ਰੂਥ
+ਸਲੇਮਾਨ_ਦਾ_ਗੀਤ
+ਉਪਦੇਸ਼ਕ
+ਵਿਰਲਾਪ
+ਅਸਤਰ
+ਦਾਨੀਏਲ
+ਅਜ਼ਰਾ
+ਨਹਮਯਾਹ
+1_ਇਤਹਾਸ
+2_ਇਤਹਾਸ