+++ tf/2021saved/book@tr.tf
@@ -1,0 +1,55 @@
+@node
+@author=Eep Talstra Centre for Bible and Computer
+@dataset=BHSA
+@datasetName=Biblia Hebraica Stuttgartensia Amstelodamensis
+@email=shebanq@ancient-data.org
+@encoders=Dirk Roorda (TF)
+@language=Türkçe
+@languageCode=tr
+@languageEnglish=turkish
+@provenance=book names from wikipedia and other sources
+@valueType=str
+@version=2021
+@website=https://shebanq.ancient-data.org
+@writtenBy=Text-Fabric
+@dateWritten=2021-08-24T11:12:50Z
+
+426591	Yaratılış
+Mısır'dan_Çıkış
+Levililer
+Çölde_Sayım
+Yasa'nın_Tekrar
+Yeşu
+Hakimler
+1_Samuel
+2_Samuel
+1_Krallar
+2_Krallar
+Yeşaya
+Yeremya
+Hezekiel
+Hoşea
+Yoel
+Amos
+Ovadya
+Yunus
+Mika
+Nahum
+Habakkuk
+Sefanya
+Hagay
+Zekeriya
+Malaki
+Mezmurlar
+Eyüp
+Süleyman'ın_Özdeyişleri
+Rut
+Ezgiler_Ezgisi
+Vaiz
+Ağıtlar
+Ester
+Daniel
+Ezra
+Nehemya
+1_Tarihler
+2_Tarihler