+++ tf/2021saved/book@ur.tf
@@ -1,0 +1,55 @@
+@node
+@author=Eep Talstra Centre for Bible and Computer
+@dataset=BHSA
+@datasetName=Biblia Hebraica Stuttgartensia Amstelodamensis
+@email=shebanq@ancient-data.org
+@encoders=Dirk Roorda (TF)
+@language=اُردُو
+@languageCode=ur
+@languageEnglish=urdu
+@provenance=book names from wikipedia and other sources
+@valueType=str
+@version=2021
+@website=https://shebanq.ancient-data.org
+@writtenBy=Text-Fabric
+@dateWritten=2021-08-24T11:12:50Z
+
+426591	پیدائش
+خروج
+احبار
+گنتی
+استثناء
+یشوع
+قضاة
+اوّل_سموئیل
+دوم_سموئیل
+اوّل_سلاطین
+دوم_سلاطین
+یسعیاہ
+یرمیاہ
+حزقی_ایل
+ہوسیع
+یوایل
+عاموس
+عبدیاہ
+یونس
+میکاہ
+نا_حُوم
+حبقّوق
+صفنیاہ
+حجيّ
+زکریاہ
+ملاکی
+زبُور
+ایّوب
+امثال
+روت
+غزل_الغزلات
+واعظ
+نوحہ
+ایستر
+دانیال
+عز_را
+نحمیاہ
+اوّل_تواریخ
+دوم_تو_اریخ